[0x12-0] Implement the print_list function
In this commit, I've added the source code for the `print_list` function, which is designed to print all the elements of a list_t list. The function takes a pointer to the list and uses a loop to traverse the list and print each element. If a string element is NULL, it prints "[0] (nil)" as specified in the task. The function returns the number of nodes in the list.

Compile command:
gcc -Wall -pedantic -Werror -Wextra -std=gnu89 0-main.c 0-print_list.c -o a

Function Prototype:
size_t print_list(const list_t *h);

File Name: 0-print_list.c